# x-log-policy-demo/src/main/java/io/xlogpolicy/demo/CustomerController.java
package io.xlogpolicy.demo;

import io.xlogpolicy.demo.model.Customer;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

/**
 * The API itself is unaffected by the policies: responses carry the real data. Only what is written to
 * the log is redacted.
 */
@RestController
@RequestMapping("/customers")
public class CustomerController {

    private final CustomerService service;

    public CustomerController(CustomerService service) {
        this.service = service;
    }

    @PostMapping
    public ResponseEntity<Customer> create(@RequestBody Customer customer) {
        return ResponseEntity.status(HttpStatus.CREATED).body(service.store(customer));
    }

    @GetMapping("/{customerId}")
    public ResponseEntity<Customer> get(@PathVariable("customerId") String customerId) {
        return service.find(customerId)
                .map(ResponseEntity::ok)
                .orElseGet(() -> ResponseEntity.notFound().build());
    }
}